+++ src/components/Dropdown/Dropdown.scss
@@ -1,4 +1,4 @@
-.MultiActionButton {
+.Dropdown {
     &__list {
         background-color: var(--color-background);
         border-radius: var(--border-radius);
