+++ src/components/Dropdown/index.ts
@@ -1,0 +1,1 @@
+export { Dropdown } from './Dropdown';
