+++ src/components/List/ListItem/ListItem.tsx
@@ -19,7 +19,7 @@ export interface Props extends Omit<React.HTMLAttributes<HTMLLIElement>, 'onClic
     highlightContext?: Context | 'default';
     /** Ref to access the li element */
     ref?: React.RefObject<HTMLLIElement>;
-    /** Item identifier is used in {@link MultiActionButton} to select/navigate through children */
+    /** Item identifier is used in {@link Dropdown} to select/navigate through children */
     value?: unknown;
 }
