--- src/components/MultiActionButton/index.ts
@@ -1,1 +1,0 @@
-export { MultiActionButton } from './MultiActionButton';
