+++ src/index.ts
@@ -33,13 +33,13 @@ export { Alert } from './components/Alert';
 export * from './components/BulkActionsToolbar';
 export { ButtonGroup } from './components/ButtonGroup';
 export { Checkbox } from './components/Checkbox';
+export { Dropdown } from './components/Dropdown';
 export { Footer } from './components/Footer';
 export { Header } from './components/Header';
 export { Field } from './components/Field';
 export { FieldWithValidation } from './components/FieldWithValidation';
 export { FieldWrapper } from './components/FieldWrapper';
 export { LoadingSpinner } from './components/LoadingSpinner';
-export { MultiActionButton } from './components/MultiActionButton';
 export { LocationCard } from './components/LocationCard';
 export { NavBar, NavItem } from './components/Navigation';
 export { Pagination } from './components/Pagination';
