+++ stories/index.ts
@@ -30,6 +30,7 @@ import './Alert';
 import './BulkActionsToolbar';
 import './ButtonGroup';
 import './Checkbox';
+import './Dropdown';
 import './Field';
 import './FieldWithValidation';
 import './FieldWrapper';
@@ -37,7 +38,6 @@ import './Footer';
 import './Header';
 import './LoadingSpinner';
 import './LocationCard';
-import './MultiActionButton';
 import './Navigation';
 import './Pagination';
 import './Pill';
